performance improvement
We now check if the input is empty or is just indicating that a key was released _before_ calling to kbd_char.
This way we improve performance as we don't call kbd_char on empty string, etc.
I also decided to change `err` to `out` as empty string is not an error.