Important
On Modern non-VM computers it might not work, since they might use Message-Signaled Interrupts instead. This works on virtual-machine that simulates the 8042 chip. 8042 chip was in charge of converting between the signals from the keyboard, and the signals to the CPU. So for example it will send an interrupt when it noticed a key was pressed.